Create settings for separate commit message files
The new .gitignore-global adds an ignore pattern for "next-commit.txt",
which can be used to stage commit messages or write them prior to making
the actual change in code. This should help with the idea that a commit
is a single thought by creating the intention before writing the code.